Refactor config update logic
