+++ backend/routes/shops.js
@@ -20,13 +20,7 @@ const {
 } = require('./_auth')
 const { createSeller } = require('../utils/sellers')
 const { getConfig, setConfig } = require('../utils/encryptedConfig')
-const {
-  createShop,
-  getShopDataUrl,
-  getShopPublicUrl,
-  getPublicUrl,
-  getDataUrl
-} = require('../utils/shop')
+const { createShop, getPublicUrl, getDataUrl } = require('../utils/shop')
 const genPGP = require('../utils/pgp')
 const get = require('lodash/get')
 const set = require('lodash/set')
@@ -38,29 +32,21 @@ const formidable = require('formidable')
 const https = require('https')
 const http = require('http')
 const mv = require('mv')
-const path = require('path')
 const { isHexPrefixed, addHexPrefix } = require('ethereumjs-util')
 
 const { configureShopDNS, deployShop } = require('../utils/deployShop')
 const { DSHOP_CACHE } = require('../utils/const')
 const { isPublicDNSName } = require('../utils/dns')
 const { getLogger } = require('../utils/logger')
-const {
-  deregisterPrintfulWebhook,
-  registerPrintfulWebhook
-} = require('../utils/printful')
 
 const printfulSyncProcessor = require('../queues/printfulSyncProcessor')
-
-const paypalUtils = require('../utils/paypal')
 
 const log = getLogger('routes.shops')
 
 const dayjs = require('dayjs')
 const { readProductsFile } = require('../utils/products')
 
-const stripeUtils = require('../utils/stripe')
-const { validateStripeKeys } = require('@origin/utils/stripe')
+const { updateShopConfig } = require('../logic/shop/config')
 
 async function tryDataDir(dataDir) {
   const hasDir = fs.existsSync(`${DSHOP_CACHE}/${dataDir}`)
@@ -731,314 +717,23 @@ module.exports = function (router) {
     }
   )
 
-  async function movePaymentMethodImages(paymentMethods, dataDir) {
-    const out = []
-    const tmpDir = path.resolve(`${DSHOP_CACHE}/${dataDir}/data/__tmp`)
-
-    for (const m of paymentMethods) {
-      const imagePath = m.qrImage
-      if (imagePath && imagePath.includes('/__tmp/')) {
-        const fileName = imagePath.split('/__tmp/', 2)[1]
-        const tmpFilePath = `${tmpDir}/${fileName}`
-
-        const targetDir = path.resolve(`${DSHOP_CACHE}/${dataDir}/data/uploads`)
-        if (!fs.existsSync(targetDir)) {
-          fs.mkdirSync(targetDir, { recursive: true })
-        }
-        const targetPath = `${targetDir}/${fileName}`
-
-        const movedFileName = await new Promise((resolve) => {
-          mv(tmpFilePath, targetPath, (err) => {
-            if (err) {
-              console.error(`Couldn't move file`, tmpFilePath, targetPath, err)
-            }
-            resolve(fileName)
-          })
-        })
-
-        out.push({
-          ...m,
-          qrImage: movedFileName
-        })
-      } else {
-        out.push(m)
-      }
-    }
-
-    return out
-  }
-
+  /**
+   * Updates a shop configuration.
+   *
+   */
   router.put(
     '/shop/config',
     authSellerAndShop,
     authRole('admin'),
     async (req, res) => {
-      // Load the network config.
-      const network = await Network.findOne({ where: { active: true } })
-      const netConfig = getConfig(network.config)
-
-      const jsonConfig = pick(
-        req.body,
-        'currency',
-        'metaDescription',
-        'css',
-        'emailSubject',
-        'cartSummaryNote',
-        'byline',
-        'discountCodes',
-        'stripe',
-        'stripeKey',
-        'title',
-        'fullTitle',
-        'facebook',
-        'twitter',
-        'instagram',
-        'medium',
-        'youtube',
-        'about',
-        'logErrors',
-        'paypalClientId',
-        'offlinePaymentMethods',
-        'supportEmail',
-        'upholdClient',
-        'useEscrow',
-        'shippingApi',
-        'taxRates',
-        'themeId'
-      )
-      const jsonNetConfig = pick(
-        req.body,
-        'acceptedTokens',
-        'customTokens',
-        'listingId',
-        'disableCryptoPayments',
-        'walletAddress'
-      )
-      const shopId = req.shop.id
-      log.info(`Shop ${shopId} - Saving config`)
-
-      let listingId
-      if (String(req.body.listingId).match(/^[0-9]+-[0-9]+-[0-9]+$/)) {
-        listingId = req.body.listingId
-        req.shop.listingId = listingId
-      }
-
-      // Offline payments
-      if (req.body.offlinePaymentMethods) {
-        jsonConfig.offlinePaymentMethods = await movePaymentMethodImages(
-          req.body.offlinePaymentMethods,
-          req.shop.authToken
-        )
-      }
-
-      if (Object.keys(jsonConfig).length || Object.keys(jsonNetConfig).length) {
-        const configFile = `${DSHOP_CACHE}/${req.shop.authToken}/data/config.json`
-
-        if (!fs.existsSync(configFile)) {
-          return res.json({ success: false, reason: 'dir-not-found' })
-        }
-
-        try {
-          const raw = fs.readFileSync(configFile).toString()
-          const config = JSON.parse(raw)
-          const newConfig = { ...config, ...jsonConfig }
-          if (Object.keys(jsonNetConfig).length) {
-            set(newConfig, `networks.${network.networkId}`, {
-              ...get(newConfig, `networks.${network.networkId}`),
-              ...jsonNetConfig
-            })
-          }
-
-          const jsonStr = JSON.stringify(newConfig, null, 2)
-          fs.writeFileSync(configFile, jsonStr)
-        } catch (e) {
-          log.error(e)
-          return res.json({ success: false })
-        }
-      }
-
-      if (req.body.about) {
-        // Update about file
-        const dataDir = `${DSHOP_CACHE}/${req.shop.authToken}/data`
-        const aboutFile = `${dataDir}/${req.body.about}`
-
-        try {
-          fs.writeFileSync(aboutFile, req.body.aboutText)
-        } catch (e) {
-          log.error('Failed to update about file', dataDir, aboutFile, e)
-          return res.json({ success: false })
-        }
-      }
-
-      const additionalOpts = {}
-
-      // Check the hostname picked by the merchant is available.
-      const existingConfig = getConfig(req.shop.config)
-      if (req.body.hostname) {
-        const hostname = kebabCase(req.body.hostname)
-        const existingShops = await Shop.findAll({
-          where: { hostname, [Sequelize.Op.not]: { id: shopId } }
+      const args = { shop: req.shop, data: req.body }
+      const result = await updateShopConfig(args)
+      if (!result.success) {
+        return res.status(400).json({
+          success: false,
+          ...pick(result, ['reason', 'field', 'message'])
         })
-        if (existingShops.length) {
-          return res.json({
-            success: false,
-            reason: 'Name unavailable',
-            field: 'hostname'
-          })
-        }
-        req.shop.hostname = hostname
-
-        // Unless explicity set by the UI, update the public and dataUrl in the
-        // shop's config to reflect a possible change to the hostname.
-        if (!req.body.publicUrl) {
-          additionalOpts.publicUrl = getShopPublicUrl(req.shop, netConfig)
-        }
-        if (!req.body.dataUrl) {
-          additionalOpts.dataUrl = getShopDataUrl(req.shop, netConfig)
-        }
       }
-      if (req.body.fullTitle) {
-        req.shop.name = req.body.fullTitle
-      }
-
-      // Configure Stripe webhooks
-      if (req.body.stripe === false) {
-        log.info(`Shop ${shopId} - Deregistering Stripe webhook`)
-        await stripeUtils.deregisterWebhooks(req.shop, existingConfig)
-        additionalOpts.stripeWebhookSecret = ''
-        additionalOpts.stripeWebhookHost = ''
-        additionalOpts.stripeBackend = ''
-      } else if (req.body.stripe) {
-        log.info(`Shop ${shopId} - Registering Stripe webhook`)
-
-        const validKeys = validateStripeKeys({
-          publishableKey: req.body.stripeKey,
-          secretKey: req.body.stripeBackend
-        })
-
-        if (!validKeys) {
-          return res.json({
-            success: false,
-            reason: 'Invalid Stripe credentials'
-          })
-        }
-
-        const { secret } = await stripeUtils.registerWebhooks(
-          req.shop,
-          req.body,
-          existingConfig,
-          req.body.stripeWebhookHost || netConfig.backendUrl
-        )
-        additionalOpts.stripeWebhookSecret = secret
-      }
-
-      // Configure Printful webhooks
-      if (req.body.printful) {
-        log.info(`Shop ${shopId} - Registering Printful webhook`)
-        const printfulWebhookSecret = await registerPrintfulWebhook(
-          shopId,
-          {
-            ...existingConfig,
-            ...req.body
-          },
-          netConfig.backendUrl
-        )
-
-        additionalOpts.printfulWebhookSecret = printfulWebhookSecret
-      } else if (existingConfig.printful && req.body.printful === false) {
-        log.info(`Shop ${shopId} - Deregistering Printful webhook`)
-        await deregisterPrintfulWebhook(shopId, existingConfig)
-        additionalOpts.printfulWebhookSecret = ''
-      }
-
-      // Duplicate `offlinePaymentMethods` to encrypted config
-      if (jsonConfig.offlinePaymentMethods) {
-        additionalOpts.offlinePaymentMethods = jsonConfig.offlinePaymentMethods
-      }
-
-      // PayPal
-      if (req.body.paypal) {
-        const client = paypalUtils.getClient(
-          netConfig.paypalEnvironment,
-          req.body.paypalClientId,
-          req.body.paypalClientSecret
-        )
-        const valid = await paypalUtils.validateCredentials(client)
-        if (!valid) {
-          return res.json({
-            success: false,
-            reason: 'Invalid PayPal credentials'
-          })
-        }
-
-        log.info(`Shop ${shopId} - Registering PayPal webhook`)
-        if (existingConfig.paypal && existingConfig.paypalWebhookId) {
-          await paypalUtils.deregisterWebhook(shopId, existingConfig, netConfig)
-        }
-        const result = await paypalUtils.registerWebhooks(
-          shopId,
-          {
-            ...existingConfig,
-            ...req.body
-          },
-          netConfig
-        )
-        additionalOpts.paypalWebhookId = result.webhookId
-      } else if (existingConfig.paypal && req.body.paypal === false) {
-        await paypalUtils.deregisterWebhook(shopId, existingConfig, netConfig)
-        additionalOpts.paypalWebhookId = null
-      }
-
-      // Save the config in the DB.
-      log.info(`Shop ${shopId} - Saving config in the DB.`)
-      const shopConfigFields = pick(
-        { ...existingConfig, ...req.body, ...additionalOpts },
-        'awsAccessKey',
-        'awsAccessSecret',
-        'awsRegion',
-        'bigQueryCredentials',
-        'bigQueryTable',
-        'dataUrl',
-        'deliveryApi',
-        'email',
-        'emailSubject',
-        'hostname',
-        'listener',
-        'mailgunSmtpLogin',
-        'mailgunSmtpPassword',
-        'mailgunSmtpPort',
-        'mailgunSmtpServer',
-        'offlinePaymentMethods',
-        'password',
-        'paypal',
-        'paypalClientId',
-        'paypalClientSecret',
-        'paypalWebhookHost',
-        'paypalWebhookId',
-        'pgpPrivateKey',
-        'pgpPrivateKeyPass',
-        'pgpPublicKey',
-        'printful',
-        'printfulAutoFulfill',
-        'processingTime',
-        'publicUrl',
-        'sendgridApiKey',
-        'sendgridPassword',
-        'sendgridUsername',
-        'shippingFrom',
-        'stripeBackend',
-        'stripeWebhookSecret',
-        'stripeWebhookHost',
-        'supportEmail',
-        'upholdApi',
-        'upholdClient',
-        'upholdSecret',
-        'web3Pk'
-      )
-      const newConfig = setConfig(shopConfigFields, req.shop.config)
-      req.shop.config = newConfig
-      req.shop.hasChanges = true
-      await req.shop.save()
       return res.json({ success: true })
     }
   )
